fix(cubestore-driver): Download x86-darwin for arm64-apple (for Rosetta2)
