+++ rust/js-wrapper/index.ts
@@ -1,1 +1,2 @@
 export * from './process';
+export { isCubeStoreSupported } from './utils';
