+++ rust/js-wrapper/utils.ts
@@ -1,5 +1,6 @@
 import colors from '@oclif/color';
 import { spawnSync } from 'child_process';
+import process from 'process';
 
 export const displayWarning = (message: string) => {
   console.log(`${colors.yellow('Warning.')} ${message}`);
@@ -45,4 +46,38 @@ export function detectLibc() {
   displayWarning('Unable to detect what host library is used as libc, continue with gnu');
 
   return 'gnu';
+}
+
+export function getTarget(): string {
+  if (process.arch === 'x64') {
+    switch (process.platform) {
+      case 'win32':
+        return 'x86_64-pc-windows-gnu';
+      case 'linux':
+        return `x86_64-unknown-linux-${detectLibc()}`;
+      case 'darwin':
+        return 'x86_64-apple-darwin';
+      default:
+        throw new Error(
+          `You are using ${process.env} platform which is not supported by Cube Store`,
+        );
+    }
+  }
+
+  if (process.arch === 'arm64' && process.platform === 'darwin') {
+    // Rosetta 2 is required
+    return 'x86_64-apple-darwin';
+  }
+
+  throw new Error(
+    `You are using ${process.arch} architecture on ${process.platform} platform which is not supported by Cube Store`,
+  );
+}
+
+export function isCubeStoreSupported(): boolean {
+  if (process.arch === 'x64') {
+    return ['win32', 'darwin', 'linux'].includes(process.platform);
+  }
+
+  return process.arch === 'arm64' && process.platform === 'darwin';
 }
